Added 3 Hooks
Added hook for forumdisplay_get_threads which is just before the database query that fetches all the threads.  
Added hook for forumdisplay_thread which is just when it starts cycling through each of the threads in case you want to do anything special with a specific thread.  
Added hook for forumdisplay_threadlist which is just before it evaluates the recentthread template for the thread.